python/requirements.txt
fastapi
uvicorn
shioaji
pandas
python-dotenv
pydantic